Baseline_models(Others)/requirements.txt
absl-py==1.4.0
aiohttp==3.8.1
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
blinker==1.5
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.4
colorlog==4.7.2
cryptography==3.4.8
decorator==5.1.1
distlib==0.3.6
docker-pycreds==0.4.0
entrypoints==0.4
filelock==3.10.7
frozenlist==1.3.3
future==0.18.2
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.1
google-auth-oauthlib==0.4.6
grpcio==1.42.0
hyperopt==0.2.7
idna==3.4
importlib-metadata==4.11.4
importlib-resources==5.12.0
ipykernel==5.5.5
ipython==7.33.0
ipython-genutils==0.2.0
jedi==0.18.1
joblib==1.2.0
jsonschema==4.17.3
jupyter-client==5.3.4
jupyter-core==4.5.0
Markdown==3.4.3
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
msgpack==1.0.5
multidict==6.0.2
networkx==2.6.3
numexpr==2.8.4
numpy==1.21.5
oauthlib==3.2.2
packaging==23.0
pandas==1.3.5
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pip==22.3.1
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
plotly==5.14.0
prompt-toolkit==3.0.36
protobuf==3.15.8
psutil==5.9.4
ptyprocess==0.7.0
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
Pygments==2.11.2
PyJWT==2.6.0
pyOpenSSL==20.0.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
pyu2f==0.1.5
PyYAML==6.0
pyzmq==23.2.0
ray==2.3.1
recbole==1.1.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.18.0
setproctitle==1.3.2
setuptools==65.6.3
six==1.16.0
smmap==5.0.0
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
torch==1.12.1
tornado==6.2
tqdm==4.65.0
traitlets==5.7.1
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.21.0
wandb==0.14.0
wcwidth==0.2.5
Werkzeug==2.1.2
wheel==0.38.4
yarl==1.7.2
zipp==3.15.0
